Add tests on proto3 optional fields
Let's add some tests on a message that contains proto3 optional
fields.

As I did not want to affect the compilation of all the test
protocol buffers with the --experimental_allow_proto3_optional
flag, compile the one .proto file intended to test this path
separately. Once the feature is no longer experimental (and
behind a flag), this .proto can be compiled along with the other
test fixtures.

Add a TODO on the single test that will be updated when we add
support for proto3 optional fields in the next commit.