Implement optional semantics for proto3
This commit adds support for the proto3 optional keyword.

Proto3 optional fields will have the label "optional" added,
to match the behaviour of proto2 optional fields.

Fields in proto3 that are not marked as optional are unaffected
and will continue to have no label (unless they are repeated, in
which case they will continue to have the "repeated" label).