Publish to DockerHub when a new tag is created
The publish to DockerHub appears to have been broken for a while now.
This attempts to address the issue by adding a new GitHub action to
handle pushing to DockerHub if the current ref is a tag.

This did involve some changes to the Dockerfile. I've tested the result
and updated the `docker_test` target accordingly.