Issue #357 Add extra steps needed for official Maven Central release.
The previous commits managed to upload the artifacts to the staging area, however these could not pass the review due to missing requirements.
This required more information to be provided in the POM file, and the releases signed.

This commit adds the information and signing.
The signing is done by GPG and requires 2 new secrets to be provided:
* GPG_SIGNING_KEY: The GPG key in armoured form.
* GPG_SIGNING_PASSWORD: The password for signing the artifact.

If a GPG key does not already exist, this guide can be followed: https://docs.github.com/en/authentication/managing-commit-signature-verification/generating-a-new-gpg-key
The value needed for the GPG_SIGNING_KEY secret can be found by:
* Finding the key name using `gpg --list-keys`
* Outputting the key using `gpg --armor --export-secret-key [name] > ~/my-key.asc`
* Copying the contents of that text file to the secret.

Once a release is done, the aritfacts will appear in the staging repository here: https://s01.oss.sonatype.org/#stagingRepositories
This staging repository can be "closed" which will execute the checks.
If all checks pass (they have in my experiments) the "Release" workflow will become availble.

The only strangeness I encountered was that originally only 2 of the releases were getting signed.
This appears to be an issue with publishing files of the same name.
I have worked around the issue by creating a copy of the files under names which include their architecture.
After I did this the signing worked correctly.